Fix: Add sidebar to all slides
This commit adds a sidebar to all slides in the presentation. It does this by using a Lua filter to insert the contents of a `sidebar.html` file into the `<body>` of the presentation.

The `_extension.yml` file has been modified to:
- Use the Lua filter.
- Read the contents of `sidebar.html` into a variable.
- Pass the variable to the Lua filter.
- Add `sidebar.html` and the lua filter to the `resources` list.

The `sidebar.html` file contains the HTML for the sidebar.

The `example.qmd` file has been modified to use only the `fnal_theme` extension.